requirements.txt
# This file may be used to create an environment
python-telegram-bot==12.8
PyYAML==5.3.1
requests==2.24.0
pandas==1.1.1
pytest==6.0.1